+++ src/Rendering/Texture.cpp
@@ -112,7 +112,7 @@ void omp::Texture::createImage()
     uint32_t size = static_cast<uint32_t>(m_TextureSource->getSize());
     uint32_t height = static_cast<uint32_t>(m_TextureSource->getHeight());
     uint32_t width = static_cast<uint32_t>(m_TextureSource->getWidth());
-    uint32_t mip_levels = static_cast<uint32_t>(m_TextureSource->getMipLevels());
+    uint32_t mip_levels = m_TextureSource->getMipLevels();
     m_VulkanContext.lock()->createImage(size,
                                         height,
                                         width, VK_FORMAT_R8G8B8A8_SRGB,
@@ -131,7 +131,7 @@ void omp::Texture::createImage()
                                               width,
                                               height);
 
-    m_VulkanContext.lock()->generateMipmaps(m_TextureImage, VK_FORMAT_R8G8B8A8_SRGB, width, height, mip_levels);
+    m_VulkanContext.lock()->generateMipmaps(m_TextureImage, VK_FORMAT_R8G8B8A8_SRGB, static_cast<int32_t>(width), static_cast<int32_t>(height), mip_levels);
 
     vkDestroyBuffer(m_VulkanContext.lock()->logical_device, staging_buffer, nullptr);
     vkFreeMemory(m_VulkanContext.lock()->logical_device, staging_buffer_memory, nullptr);
